Makefile: return error if a check fails
Make target pythoncheck (and codecheck) did not fail if the last tested
file was error-free. Now we set RETURN=1 if only one check fails.

While at it also created separate targets for flake8, black and isort,
but also added --keep-going option when running 'make codecheck' so that all
targets are run.

Added @ in front of each command to hide the command which makes the
output more readable.

Spellcheck SC1117 (Backslash is literal in "\n") is no longer emitted
(with shellcheck after v0.5), thats why it is removed.

Please note, that spellintian does not return an error code if an
incorrect spelling is found, that is why this ugly hack had to be
implemented.