grml-resolution: Quick 'n dirty xrandr output parser fix
The xrandr output parser of grml-resolution seems to be broken since at
least Grml 2014.11: grml-resolution does not list possible resolutions
but its corresponding refresh rates.

It seems like once upon a time xrandr had a 'size-index' in front of
each resolution. This is not the case anymore and there does not seem to
be an option to add that behaviour back to xrandr.
So, CURRENT_NUM did not receive the 'size-index' of the current
resolution, therefor I dropped that variable which was used to detect if
the user had chosen the current resolution (nothing bad happens if
xrandr sets the current resolution).

We now list the possible resolutions and refresh rates, separated by an
underscore because `dialog` does not like spaces.
xrandr -s works fine with the chosen 'index' from dialog.

Closes: grml/grml-scripts#14